Reorganise LocalImagesViewController class
