+++ piwigo/Upload/Pick Local Images/LocalImagesViewController.storyboard
@@ -152,7 +152,7 @@
                                         </segments>
                                         <color key="tintColor" red="1" green="0.46666666670000001" blue="0.0039215686269999999" alpha="1" colorSpace="calibratedRGB"/>
                                         <connections>
-                                            <action selector="didChangeSortOption:" destination="0bN-Jj-8Z5" eventType="valueChanged" id="XCZ-g0-R6J"/>
+                                            <action selector="didChangeGroupOption:" destination="0bN-Jj-8Z5" eventType="valueChanged" id="DUC-rV-Zya"/>
                                         </connections>
                                     </segmentedControl>
                                 </subviews>
